goofy

+++ Backend/backend/src/main/java/recipeapp/Recipes/Recipe.java
@@ -57,7 +57,7 @@ public class Recipe {
 
 
     /** The rating of the recipe given by users. */
-    private int rating = 0;
+    private double rating = 0;
     private int ratingCount = 0;
     private int totalRating = 0;
 }
